apps/processor/microservices/units/ms-atm-ai/pyproject.toml
[project]
name = "ms-atm-ai"
version = "0.1.0"
description = "AI Gateway microservice – semantic analysis, embeddings, and data cleaning via LiteLLM."
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.62",
    "grpcio-tools>=1.62",
    "grpcio-health-checking>=1.62",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tiktoken>=0.6",
    "opentelemetry-api>=1.23",
    "opentelemetry-sdk>=1.23",
    "opentelemetry-instrumentation-grpc>=0.44b0",
    "opentelemetry-exporter-otlp>=1.23",
    "dapr>=1.13",
    "dapr-ext-grpc>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 120
